--- client/ExpoParty/README.md
@@ -1,3 +1,0 @@
-# expo-pary
-
-### A beta library implementing Game Center / XBox Live features
