--- client/Manager/Manager.js
@@ -1,9 +1,0 @@
-class Manager {
-  get assets() {
-    return {};
-  }
-
-  async setupAsync() {}
-}
-
-export default Manager;
