+++ client/components/tabBarIcon.js
